refactor: broke down coordinator class to improve maintainability
